moved some logic and fixed some bugs
