+++ components/upload/UploadPanel.tsx
@@ -5,9 +5,9 @@ import DropdownButton from "@/components/util/DropdownButton";
 import DropdownItems from "@/components/util/DropdownItems";
 import FileInputArea from "@/components/util/FileInputArea";
 import SignMessageButton from "@/components/util/SignMessageButton";
-import { uploadFile, whitelistFromFile } from "api/uploadFile";
+import { uploadFile, whitelistFromFile } from "@/api/upload/uploadFile";
 import SwitchButton from "./SwitchButton";
-import { uploadSingle } from "api/uploadSingle";
+import { uploadSingle } from "api/upload/uploadSingle";
 import FormInputText from "@/components/util/FormInputText";
 import ErrorMessage from "@/components/util/ErrorMessage";
 import { IResponseMessage } from "@/models/IResponseMessage";
